Fix: 수정페이지 폴더 구조 변경
