+++ src/Pages/CreatePotPage/Components/DetailPage/EditDeleteModal.tsx
@@ -88,12 +88,6 @@ const EditDeleteModal = ({ postId, isOpen, onClose }: EditDeleteModalProps) => {
     );
 };
 
-// EditDeleteModal.propTypes = {
-//     isOpen: PropTypes.bool.isRequired,
-//     onClose: PropTypes.func.isRequired,
-//     children: PropTypes.node.isRequired,
-// };
-
 export default EditDeleteModal;
 
 const BottomSheetContainer = styled.div`
