+++ src/Pages/UpdatePage/UpdateDescription.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent } from 'react';
-import usePostDataStore from '../../../../state/store/PostDataStore';
-import * as S from '../../style';
+import usePostDataStore from '../../state/store/PostDataStore';
+import * as S from '../CreatePotPage/style';
 
 function CreateDescription() {
     const { data, setPostData } = usePostDataStore();
